content-planner/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
PyJWT>=2.8.0
httpx>=0.24.0
prometheus-client>=0.17.0
slowapi>=0.1.9
